+++ services/automation-scripting-service/src/main/resources/application.yml
@@ -28,6 +28,7 @@ server:
     min-response-size: 1024
 
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
