+++ services/game-logic-service/src/main/resources/application-prod.yml
@@ -16,6 +16,7 @@ management:
         enabled: true
 
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
