+++ services/game-logic-service/src/main/resources/application.yml
@@ -17,6 +17,7 @@ management:
         enabled: true
 
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
