+++ services/game-session-service/src/main/resources/application.yml
@@ -31,6 +31,7 @@ server:
     mime-types: application/json,application/xml,text/html,text/xml,text/plain
     min-response-size: 1024
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
