+++ services/logging-admin-service/src/main/resources/application-prod.yml
@@ -22,6 +22,7 @@ server:
     min-response-size: 1024
 
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
