+++ services/logging-admin-service/src/main/resources/application.yml
@@ -23,6 +23,7 @@ server:
     min-response-size: 1024
 
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
