+++ services/social-groups-service/src/main/resources/application-prod.yml
@@ -51,6 +51,7 @@ server:
     mime-types: application/json,application/xml,text/html,text/xml,text/plain
     min-response-size: 1024
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
