+++ services/social-groups-service/src/main/resources/application.yml
@@ -50,6 +50,7 @@ server:
     mime-types: application/json,application/xml,text/html,text/xml,text/plain
     min-response-size: 1024
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
