+++ services/world-management-service/src/main/resources/application-prod.yml
@@ -29,6 +29,7 @@ server:
     min-response-size: 1024
 
 grpc:
+  shutdownGrace: 30
   server:
     port: 6565
     enableKeepAlive: true
